An item is loaded from the cache when is locked
Fixing issue when item is loaded from the cache and is eventually locked. If the item is locked, then we load an old copy, because the locking worker may make changes to the item during the lock. With this change we are avoiding retrieving an old copy of the item.